+++ Software/util/RNG.h
@@ -20,7 +20,9 @@
 
 #include <stdint.h>
 
-void RNG_init();
+void RNG_seed();
+
+uint8_t RNG_nextByte();
 
 uint8_t RNG_nextChar();