+++ CHANGELOG.md
@@ -8,6 +8,13 @@ For changes to the hardware of the project please jump to the # Hardware section
 ### [UNRELEASED] - [UNRELEASED]
 #### Adding
 - BOM
+- User Manual
+
+#### Changing
+- Visually correct world line rolling
+
+#### Fixing
+- Display errors on tube3 and tube6
 
 
 ### [0.1.0a] - 2015-07-23